backend/requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# NLP & Sentiment Analysis
vaderSentiment==3.3.2
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
scikit-learn==1.3.2
shap==0.43.0

# Financial Data
yfinance==0.2.32
alpha-vantage==2.3.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Async Task Queue
celery==5.3.4
redis==5.0.1

# AWS Services
boto3==1.34.10

# API & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Statistics & Math
scipy==1.11.4
statsmodels==0.14.0

# Utilities
python-dotenv==1.0.0
schedule==1.2.0

# Monitoring & Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Payment Processing
stripe==7.8.0

# Development
black==23.12.0
flake8==6.1.0

